+++ docs/pages/hotkeys/index.js
@@ -1,5 +1,5 @@
 import Markdown from '../../libs/markdown/';
-
+import './index.less';
 export default class App extends Markdown {
-  
+
 }