+++ src/date-picker/DatePanelBody.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Component, PropTypes, formatDate, isDate } from '../utils/';
-import { fillUpDays } from './utils';
 import DatePanelHead from './DatePanelHead';
 import DatePanelMode from './DatePanelMode';
 import TimePickerSpinner from './TimePickerSpinner';
+import DatePanelBodyDay from './DatePanelBodyDay';
 
 export default class DatePanelBody extends Component {
   constructor(props) {
@@ -113,12 +113,18 @@ export default class DatePanelBody extends Component {
     shortcut.onClick();
   }
   render() {
-    const { prefixCls, weekLabel, format, onPicked, shortcutinline, showTime, renderDate, shortcutClassName, shortcuts } = this.props;
+    const { prefixCls, format, weekLabel, onPicked, shortcutinline, showTime, renderDate, shortcutClassName, disabledDate, shortcuts } = this.props;
     const { value, labelToday, selectDate, selectTime, selectYear, selectMonth, labelTimeVisible } = this.state;
     const datePanel = isDate(value) ? new Date(value) : new Date();
+
     const headerProps = {
-      prefixCls, value: datePanel, defaultValue: this.props.value, selectYear, selectMonth, selectDate, onPicked,
+      prefixCls, value: datePanel, defaultValue: this.props.value, selectYear, selectMonth, selectDate, onPicked, disabledDate,
     };
+    if (selectDate && isDate(selectDate)) {
+      headerProps.selectDate = new Date(selectDate);
+    } else {
+      headerProps.selectDate = null;
+    }
 
     const DatePanelHeadLabel = (
       <DatePanelHead {...headerProps}
@@ -162,40 +168,16 @@ export default class DatePanelBody extends Component {
           <TimePickerSpinner {...timeProps} />
         )}
         {DatePanelHeadLabel}
-        <div className={`${prefixCls}-week`}>
-          {weekLabel.map((label, idx) => {
-            return (
-              <span key={idx} className={this.classNames({ end: idx === 0 || idx === 6 })} >
-                {label}
-              </span>
-            );
-          })}
-        </div>
-        <div className={`${prefixCls}-days`}>
-          {fillUpDays(datePanel, format, new Date(selectDate)).map((item, idx) => {
-            if (renderDate) {
-              const child = renderDate(item, item.selectDay && selectDate);
-              return React.cloneElement(child, {
-                key: idx,
-                onClick: () => this.handleClick(item),
-              });
-            }
-            return (
-              <span key={idx}
-                title={item.today ? labelToday : item.format}
-                className={this.classNames(item.className, {
-                  [`${prefixCls}-today`]: item.today,
-                  [`${prefixCls}-select-day`]: item.selectDay && selectDate,
-                  [`${prefixCls}-sun`]: item.week === 0,
-                  [`${prefixCls}-sat`]: item.week === 6,
-                })}
-                onClick={() => this.handleClick(item)}
-              >
-                {item.day}
-              </span>
-            );
-          })}
-        </div>
+        <DatePanelBodyDay
+          format={format}
+          weekLabel={weekLabel}
+          selectDate={headerProps.selectDate}
+          disabledDate={disabledDate}
+          date={datePanel}
+          renderDate={renderDate}
+          labelToday={labelToday}
+          onClick={this.handleClick.bind(this)}
+        />
         {
           shortcuts && Array.isArray(shortcuts) && (
             <div className={
@@ -233,6 +215,7 @@ DatePanelBody.propTypes = {
   allowClear: PropTypes.bool,
   onPicked: PropTypes.func,
   renderDate: PropTypes.func,
+  disabledDate: PropTypes.func,
   showToday: PropTypes.oneOfType([
     PropTypes.bool,
     PropTypes.node,
@@ -241,7 +224,6 @@ DatePanelBody.propTypes = {
     PropTypes.bool,
     PropTypes.object,
   ]),
-  weekLabel: PropTypes.arrayOf(PropTypes.string),
 };
 
 DatePanelBody.defaultProps = {
@@ -250,6 +232,5 @@ DatePanelBody.defaultProps = {
   showToday: false, // 是否展示“今天”按钮
   showTime: false, // 是否展示“选择时间”按钮
   prefixCls: 'w-datepicker',
-  weekLabel: ['日', '一', '二', '三', '四', '五', '六'],
   onPicked() { },
 };