+++ src/date-picker/index.js
@@ -4,7 +4,5 @@ import './style/index.less';
 
 export { default as TimeSelect } from './TimeSelect';
 export { default as TimePicker } from './TimePicker';
-// export { default as DatePicker } from './DatePicker';
-
 
 export default DatePicker;