+++ src/timestamp/__test__/Timestamp.test.js
@@ -12,7 +12,7 @@ describe('<Timestamp>', () => {
     expect(wrapper.find('.w-timestamp')).toHaveLength(1);
     expect(wrapper.type()).toEqual(Timestamp);
     expect(wrapper.at(0).prop('prefixCls')).toBe('w-timestamp');
-    expect(wrapper.at(0).prop('format')).toBe('yyyy-MM-dd hh:mm:ss');
-    expect(wrapper.html()).toContain('<span class="w-timestamp" value="Wed Nov 22 2017 02:06:01 GMT+0800 (CST)" format="yyyy-MM-dd hh:mm:ss">2017-11-22 02:06:01</span>');
+    expect(wrapper.at(0).prop('format')).toBe('Y-m-d h:i:s');
+    expect(wrapper.html()).toContain('<span class="w-timestamp" interval="1000">2017-11-22 02:06:01</span>');
   });
 });